Tidy up map decoding
